Move backendcalls package to network package
